requirements.txt
pynput
pydbus
sounddevice
faster-whisper
pyperclip
numpy
scipy
dbus-python 
PyYAML
pydantic 